% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{pin_action}
\alias{pin_action}
\title{Pin an individual action to its corresponding SHA}
\usage{
pin_action(action = "r-lib/actions/check-r-package@v2")
}
\arguments{
\item{action}{the action written with github syntax that contains at least
the github user, repo, and tag}
}
\value{
the same action with an SHA and a comment that indicates the tag
the sha belongs to (this may be different than the tag if you use a floating tag)
}
\description{
Pin an individual action to its corresponding SHA
}
\examples{
pin_action("r-lib/actions/check-r-package@v2")
}
