% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{pin_action_workflow}
\alias{pin_action_workflow}
\title{Pin a third-party action in a workflow}
\usage{
pin_action_workflow(
  action = "r-lib/actions/check-r-package@v2",
  replacement = NULL,
  workflow = ".github/workflows/R-CMD-check.yaml",
  write = FALSE
)
}
\arguments{
\item{action}{the action written with github syntax that contains at least
the github user, repo, and tag}

\item{replacement}{(optional) the replacement for the particular action. If
this is not provided (default), \code{\link[=pin_action]{pin_action()}} will be used to find the
replacement}

\item{workflow}{the path to a github workflow file}

\item{write}{if \code{TRUE}, the workflow file will be overwritten. Defaults to
\code{FALSE}, which leaves the workflow file intact}
}
\value{
a character vector of the workflow file with the actions replaced.
}
\description{
Pin a third-party action in a workflow
}
\examples{
\dontshow{if (requireNamespace("withr", silently = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pkgdown <- withr::local_tempdir()
fs::file_copy(system.file("workflows", "pkgdown.yaml", package = "pinsha"), pkgdown)
actions <- pin_find_actions(pkgdown)
actions
for (action in actions) pin_action_workflow(action, workflow = pkgdown, write = TRUE)
pin_find_actions(pkgdown)
\dontshow{\}) # examplesIf}
}
