% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{pin_find_actions}
\alias{pin_find_actions}
\title{Find GitHub actions used in a workflow file}
\usage{
pin_find_actions(
  workflow = ".github/workflows/R-CMD-check.yaml",
  include_official = FALSE
)
}
\arguments{
\item{workflow}{the path to a github workflow file}

\item{include_official}{when \code{TRUE}, official actions (that have the
organization name of \verb{actions/}) will be included in the results. Defaults
to \code{FALSE}.}
}
\value{
a character vector of zero or more actions used in your workflows
}
\description{
Find GitHub actions used in a workflow file
}
\examples{
workflows <- system.file("workflows", package = "pinsha")
pin_find_actions(workflows)
pkgdown <- system.file("workflows", "pkgdown.yaml", package = "pinsha")
pin_find_actions(pkgdown)
}
