% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinsha-package.R
\docType{package}
\name{pinsha-package}
\alias{pinsha}
\alias{pinsha-package}
\title{pinsha: Pin 'GitHub' Actions to 'SHA'}
\description{
Harden your 'GitHub' workflow security by pinning your third-party actions to the appropriate 'SHA'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zkamvar/pinsha}
  \item Report bugs at \url{https://github.com/zkamvar/pinsha/issues}
}

}
\author{
\strong{Maintainer}: Zhian N. Kamvar \email{zkamvar@gmail.com} (\href{https://orcid.org/0000-0003-1458-7108}{ORCID})

}
\keyword{internal}
