% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{pin}
\alias{pin}
\title{Automatically pin all actions to their expected hashes in your GitHub workflows}
\usage{
pin(workflows = ".github/workflows", include_official = FALSE, write = FALSE)
}
\arguments{
\item{workflows}{the directory to your github workflows}

\item{include_official}{when \code{TRUE}, official github action workflows will also be pinned. Defaults to \code{FALSE}, meaning that the official workflows will continue to use tags}

\item{write}{When \code{TRUE}, the workflows will be overwritten. Defaults to \code{FALSE}}
}
\value{
nothing. Used for its side-effect
}
\description{
Automatically pin all actions to their expected hashes in your GitHub workflows
}
\examples{
\dontshow{if (requireNamespace("withr", silently = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
tmp <- withr::local_tempdir()
workflows <- fs::path(tmp, ".github", "workflows")
fs::dir_copy(system.file("workflows", package = "pinsha"), workflows)
withr::with_dir(tmp, pin_find_actions(".github/workflows"))
withr::with_dir(tmp, pin(write = TRUE))
withr::with_dir(tmp, pin_find_actions(".github/workflows"))
\dontshow{\}) # examplesIf}
}
